# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API token authentication. Use format: \"Bearer \u003cyour-api-token\u003e\"",
        "title": "Bible Notes API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "shuvoedward@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/v1",
    "paths": {
...
        "/v1/autocomplete": {
            "get": {
                "description": "Provides intelligent autocomplete suggestions based on the search query type. Returns book name suggestions for partial book names, word matches for single words (3+ character), or verse reference for multi-word queries containing book names and references.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search",
                    "Autocomplete"
                ],
                "summary": "Autocomplete Bible search queries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query (e.g., 'mat', 'john 3', 'love')",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully returned autocomplete suggestions. Type indicates the search category: 'book' for book name matches, 'word' for word searches, or 'verse' for verse reference searches",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "autocomplete": {
                                    "type": "object",
                                    "properties": {
                                        "books": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "type": {
                                            "type": "string"
                                        },
                                        "verses": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.VerseMatch"
                                            }
                                        },
                                        "words": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.WordMatch"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Query parameter is empty or missing",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/bible/{book}/{chapter}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves the text for a specified Bible chapter or a range of verses, along with associated user-specific data (highlights and notes) if the user is logged in and activated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bible",
                    "Passages"
                ],
                "summary": "Get a Bible chapter or verse range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The name of the Bible book (e.g., Genesis)",
                        "name": "book",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The chapter number (e.g., 1)",
                        "name": "chapter",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Start verse number (must be used with evs)",
                        "name": "svs",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "End verse number (must be used with svs)",
                        "name": "evs",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved passage and user data",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "bible_notes": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.NoteResponse"
                                    }
                                },
                                "cross_ref_notes": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.NoteResponse"
                                    }
                                },
                                "highlights": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.Highlight"
                                    }
                                },
                                "passage": {
                                    "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.Passage"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters (e.g., invalid chapter or verse numbers)",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Passage not found (e.g., invalid book/chapter combination)",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/grammar/check": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Check text for grammar, spelling, and style issues using LanguageTool API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grammar"
                ],
                "summary": "Check grammar and spelling",
                "parameters": [
                    {
                        "description": "Text to check",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "text": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Grammar check results with suggestions",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "matches": {
                                    "$ref": "#/definitions/cmd_api.GrammarResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error or LanguageTool API error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/search/bible": {
            "get": {
                "description": "Performs a full-text search across Bible verses to find occurrences of the specified word or phrase. Returns paginated results with verse references and surrounding context.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search",
                    "Verses"
                ],
                "summary": "Search Bible verses by word",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search word or phrase (e.g., 'love', 'faith')",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page number (minimum: 1)",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Number of results per page (minimum: 1)",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully returned paginated verse search results",
                        "schema": {
                            "type": "object",
                            "properties": {
                                " metadata": {
                                    "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.Metadata"
                                },
                                "verses": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.VerseMatch"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters (empty query, empty metadata)",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cmd_api.GrammarMatch": {
            "type": "object",
            "properties": {
                "length": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "offset": {
                    "type": "integer"
                },
                "replacements": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "cmd_api.GrammarResponse": {
            "type": "object",
            "properties": {
                "matches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cmd_api.GrammarMatch"
                    }
                }
            }
        },
        "shuvoedward_Bible_project_internal_data.Highlight": {
            "type": "object",
            "properties": {
                "book": {
                    "type": "string"
                },
                "chapter": {
                    "type": "integer"
                },
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "end_offset": {
                    "type": "integer"
                },
                "end_verse": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "start_offset": {
                    "type": "integer"
                },
                "start_verse": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "shuvoedward_Bible_project_internal_data.LocationResponse": {
            "type": "object",
            "properties": {
                "book": {
                    "type": "string"
                },
                "chapter": {
                    "type": "integer"
                },
                "end_offset": {
                    "type": "integer"
                },
                "end_verse": {
                    "type": "integer"
                },
                "location_id": {
                    "type": "integer"
                },
                "start_offset": {
                    "type": "integer"
                },
                "start_verse": {
                    "type": "integer"
                }
            }
        },
        "shuvoedward_Bible_project_internal_data.Metadata": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "first_page": {
                    "type": "integer"
                },
                "last_page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                }
            }
        },
...
        "shuvoedward_Bible_project_internal_data.NoteResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.LocationResponse"
                },
                "note_id": {
                    "type": "integer"
                },
                "note_type": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "shuvoedward_Bible_project_internal_data.Passage": {
            "type": "object",
            "properties": {
                "book": {
                    "type": "string"
                },
                "chapter": {
                    "type": "integer"
                },
                "verses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.VerseDetail"
                    }
                }
            }
        },
...
        "shuvoedward_Bible_project_internal_data.VerseDetail": {
            "type": "object",
            "properties": {
                "number": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "verse_id": {
                    "type": "integer"
                }
            }
        },
        "shuvoedward_Bible_project_internal_data.VerseMatch": {
            "type": "object",
            "properties": {
                "book": {
                    "type": "string"
                },
                "chapter": {
                    "type": "integer"
                },
                "exactMatch": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "rank": {
                    "type": "number",
                    "format": "float64"
                },
                "snippet": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "verse": {
                    "type": "integer"
                }
            }
        },
        "shuvoedward_Bible_project_internal_data.WordMatch": {
            "type": "object",
            "properties": {
                "frequency": {
                    "type": "integer"
                },
                "lexeme": {
                    "type": "string"
                },
                "word": {
                    "type": "string"
                }
            }
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ]
}